+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 5.1.0
 
 - Realigned code base with Flutter `3.24.0`
+- Removed dependency to package `collection`
 
 ## 5.0.2